% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpod_fns.R
\name{create_gp_dockerfile}
\alias{create_gp_dockerfile}
\title{Create a gitpod ready dockerfile from a R description file}
\usage{
create_gp_dockerfile(FROM, AS, rstudio)
}
\arguments{
\item{FROM}{\code{character} FROM of the Dockerfile}

\item{AS}{\code{character} AS of the Dockerfile}

\item{rstudio}{\code{logical} if using a rstudio server image open ports}
}
\value{
A Dockerfile Object.
}
\description{
Create a gitpod ready dockerfile from a R description file
}
