% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpod_fns.R
\name{create_gp_files}
\alias{create_gp_files}
\title{Creates a docker- and .yml file from
the current R project to run in a gitpod instance
with relevant dependencies installed.}
\usage{
create_gp_files(FROM, AS = NULL, rstudio = TRUE)
}
\arguments{
\item{FROM}{\code{character} FROM of the Dockerfile}

\item{AS}{\code{character} AS of the Dockerfile}

\item{rstudio}{\code{logical} if using a rstudio server image open ports}
}
\value{
a docker- and .yml file
}
\description{
Creates a docker- and .yml file from
the current R project to run in a gitpod instance
with relevant dependencies installed.
}
