% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpod_fns.R
\name{create_gp_yml}
\alias{create_gp_yml}
\title{Create  a .gitpd.yml file for R(studio)}
\usage{
create_gp_yml(rstudio)
}
\value{
A yml file wrtitten to .gitpod.yml
}
\description{
Create  a .gitpd.yml file for R(studio)
}
